add production build
