+++ README.md
@@ -4,11 +4,22 @@ Simple Trello Design
 ## Repository
 https://github.com/shin-dev/trello-design
 
-## Contributing
+## Development
 
 ```scripts
 $ yarn install
+
+# Generate a "dist" directory and watching the src directory files.
 $ yarn run dev
+```
+
+## Release
+
+```scripts
+$ yarn install --frozen-lockfile
+
+# Generate a "dist" directory, and generate a "trello-design.zip" in it.
+$ yarn run release
 ```
 
 ## License
