+++ src/scripts/options.js
@@ -1,0 +1,4 @@
+/*
+ * © 2020 シン合同会社 - SHIN LLC
+ * https://shin-dev.com
+ */
