+++ src/scripts/popup.js
@@ -1,3 +1,7 @@
+/*
+ * © 2020 シン合同会社 - SHIN LLC
+ * https://shin-dev.com
+ */
 window.onload = () => {
   // Initialize
   const links = document.getElementsByTagName('a')
